% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors.R
\name{add_author}
\alias{add_author}
\alias{add_first_author}
\alias{add_coauthors}
\title{Add authors to the project metadata.}
\usage{
add_first_author(first, last, role = c("aut", "cre"), email = NULL,
  orcid = NULL)

add_coauthors(first, last, role = "ctb", email = NULL, orcid = NULL)
}
\arguments{
\item{first}{Given or first name.}

\item{last}{Last or family name.}

\item{role}{The role of the author. See details below.}

\item{email}{Optional. Email address.}

\item{orcid}{Optional. \href{https://orcid.org/}{ORCID} for author.}
}
\value{
Adds authors to DESCRIPTION file (the project metadata).
}
\description{
To give credit where credit is due, add these people to the DESCRIPTION file
(the project metadata) as (co-)authors. First author is the person responsible
for the Git repository, not necessarily for the scientific product.
}
\details{
For author roles:
\itemize{
\item "cre" is the one who owns and is responsible for the Git repository and
for the code.
\item "aut" is someone who has contributed substantially to the code, analysis,
and repository.
\item "ctb" is someone who has contributed non-code and non-analysis components
to the project.
}
}
\examples{
\dontrun{
# Owner of repo/primarily author who coded/analyzed data.
add_first_author("Jim", "Joe", orcid = "0456436")

# Made substantial contributions to code/analysis
add_coauthor("Jilly", "Bean", role = "aut")

# Made non-code/analysis contributions
add_coauthor("Tina", "Teller", role = "ctb")
}
}
