% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_etl}
\alias{create_etl}
\title{Create a standarized file inside the etl folder}
\usage{
create_etl(prefix = "001")
}
\arguments{
\item{prefix}{How to name your file}
}
\value{
A created \code{.R} file in the \verb{etl/} folder.
}
\description{
Create a standarized file inside the etl folder
}
\examples{
\dontrun{
create_etl('001_dw')
create_etl('002_dw_users')
}

}
