% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createDoc.R
\name{createDoc}
\alias{createDoc}
\title{Create documents}
\usage{
createDoc(file, output.dir = "./output")
}
\arguments{
\item{file}{file to convert to output formats}

\item{output.dir}{path to output folder}
}
\value{
Output document from Rmarkdown conversion.
}
\description{
Create documents (word, html) from R scripts or R Markdown files
}
\details{
This is a thin wrapper around the \code{\link[rmarkdown]{render}}
function.
}
\author{
Luke W. Johnston
}

