% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileTemplate.R
\name{fileTemplate}
\alias{fileTemplate}
\title{Create a file based on a template in the \code{prodigenr} package.}
\usage{
fileTemplate(template, output.file = NULL)
}
\arguments{
\item{template}{template file to use}

\item{output.file}{}
}
\value{
Creates a file from a template
}
\description{
File creation from a template
}
\details{
Creates a file from a template (shown in the \code{\link[prodigenr]{listTemplates}} function)
}
\author{
Luke W. Johnston
}
\seealso{
\code{\link[prodigenr]{listTemplates}} for a list of
templates available.  \code{\link[rmarkdown]{draft}} for a
similar, more developed but slightly more complex function.
}

