% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigData.R
\name{getConfigData}
\alias{getAuthor}
\alias{getBib}
\alias{getCSL}
\alias{getConfigData}
\alias{getData}
\title{Extract data from the config file}
\usage{
getBib(config.file)

getCSL(config.file)

getData(config.file)

getAuthor(config.file)
}
\arguments{
\item{config.file}{location of the config file}
}
\description{
Extract data from the config file
}
\details{
A series of functions to extract data from the configuration file
found in the parent directory.
}
\author{
Luke W. Johnston
}

