% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{include_mit_license}
\alias{include_mit_license}
\title{License the project under MIT.}
\usage{
include_mit_license(path = ".")
}
\arguments{
\item{path}{Path to the project.}
}
\value{
Creates a MIT LICENSE file.
}
\description{
License the project under MIT.
}
