% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{include_rfigshare_script}
\alias{include_rfigshare_script}
\title{Add an R script to send code to figshare.}
\usage{
include_rfigshare_script(path = ".")
}
\arguments{
\item{path}{Path to the project.}
}
\value{
Creates an rfigshare script.
}
\description{
Add an R script to send code to figshare.
}

