% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{include_strobe}
\alias{include_strobe}
\title{Include a STROBE checklist in the project.}
\usage{
include_strobe(path = ".")
}
\arguments{
\item{path}{Path to the project.}
}
\value{
Creates a file with the STROBE checklist.
}
\description{
Include a STROBE checklist in the project.
}

