% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_templates.R
\docType{data}
\name{list_templates}
\alias{list_templates}
\alias{template_list}
\title{List project templates within \pkg{prodigenr}.}
\format{An object of class \code{character} of length 4.}
\usage{
list_templates(type = "projects", package = "prodigenr")

template_list
}
\arguments{
\item{type}{Type of template (file or directory, aka project)}

\item{package}{Name of package}
}
\value{
Vector of templates available
}
\description{
Get a list of available templates in a package.
}
\examples{

list_templates('projects')
list_templates()

}
\keyword{datasets}

