% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load .RData file(s)}
\usage{
loadData(path)
}
\arguments{
\item{path}{Location of data files}
}
\description{
Search for .RData file(s) and load into the workspace
}
\author{
Luke W. Johnston
}

