% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_docs.R
\name{render_docs}
\alias{render_docs}
\title{Rstudio addin to render all files in the doc/ folder.}
\usage{
render_docs()
}
\value{
Generated rmarkdown output.
}
\description{
Rstudio addin to render all files in the doc/ folder.
}

