% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSetup.R
\name{runSetup}
\alias{runSetup}
\title{Setup a script for analysis}
\usage{
runSetup(path)
}
\arguments{
\item{path}{location of the setup files.}
}
\description{
Load setup.R and functions.R files
}
\details{
This function is a thin wrapper around \code{\link{source}}.
}
\author{
Luke W. Johnston
}

