% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{set_git_to_project}
\alias{set_git_to_project}
\title{Setup Git to the project.}
\usage{
set_git_to_project()
}
\value{
Adds Git and \code{.gitignore} file to the project.
}
\description{
Takes a lot of inspiration from the usethis \code{use_git()} function, except
it only adds Git and nothing more (doesn't commit, doesn't restart RStudio
automatically). Must run this function inside the project you created from
\code{\link[=setup_project]{setup_project()}}
}
\seealso{
\code{\link[=setup_project]{setup_project()}} for starting the project.
}
