% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup a research analysis project.}
\usage{
setup_project(name, path = getwd())
}
\arguments{
\item{name}{Name of project.}

\item{path}{Path/location of project.}
}
\value{
Folders and files created for a research project.
}
\description{
This starts the project by setting up a common folder and file infrastructure,
as well as adding some useful files to start.
}
\examples{
\dontrun{
path <- tempdir()
setup_project("DiabetesCancer", path)
}
}
