% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_list.R
\docType{data}
\name{template_list}
\alias{template_list}
\title{List project templates within \pkg{prodigenr}.}
\format{An object of class \code{character} of length 4.}
\usage{
template_list
}
\value{
Vector of templates available
}
\description{
Get a list of available templates in a package.
}
\examples{
template_list

}
\keyword{datasets}
