% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_doc.R
\name{create_doc}
\alias{create_doc}
\alias{create_manuscript}
\alias{create_poster}
\alias{create_slides}
\alias{create_abstract}
\title{Create an R Markdown document from template.}
\usage{
create_doc(type = c("manuscript", "slides", "poster", "abstract"))

create_manuscript()

create_poster()

create_slides()

create_abstract()
}
\arguments{
\item{type}{The file type (e.g. manuscript, slides, etc.)}
}
\value{
A created Rmd file.
}
\description{
Creates manuscript, slide, poster, or abstract R Markdown files into the doc folder.
}
\examples{
\dontrun{
create_manuscript()
create_poster()
create_slides()
create_abstract()
}
}
