% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileTemplate.R
\name{fileTemplate}
\alias{fileTemplate}
\title{Create a file based on a template in the \code{prodigenr} package.}
\usage{
fileTemplate(template, output.file = NULL)
}
\arguments{
\item{template}{Template file to use}

\item{output.file}{Where the template file should be created in}
}
\value{
Creates a file from a template
}
\description{
File creation from a template
}
\details{
Creates a file from a template (shown in the \code{\link[prodigenr]{listTemplates}} function)
}
\examples{
\dontrun{
## Get a list of the possible file templates
listTemplates('files')
fileTemplate('eda.R', './eda.R')
fileTemplate('eda.R', 'path/to/dir/eda.R')
}
}
\seealso{
\code{\link[prodigenr]{listTemplates}} for a list of
templates available.  \code{\link[rmarkdown]{draft}} for a
similar, more developed but slightly more complex function.
}

