% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_templates.R
\name{list_templates}
\alias{list_templates}
\title{List project templates within \pkg{prodigenr}.}
\usage{
list_templates(type = "projects", package = "prodigenr")
}
\arguments{
\item{type}{Type of template (file or directory, aka project)}

\item{package}{Name of package}
}
\value{
Vector of templates available
}
\description{
Get a list of available templates in a package.
}
\examples{

list_templates('projects')
list_templates()

}

