% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listTemplates.R
\name{listTemplates}
\alias{listTemplates}
\title{List templates within \pkg{prodigenr}.}
\usage{
listTemplates(type = c("files", "projects"), package = "prodigenr")
}
\arguments{
\item{type}{type of template (file or directory, aka project)}

\item{package}{name of package}
}
\value{
Vector of templates available
}
\description{
Get a list of available templates in a package
}
\author{
Luke W. Johnston
}

