% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodigen.R
\name{prodigen}
\alias{prodigen}
\title{Create a project directory based on a template}
\usage{
prodigen(proj.type, proj.name = NULL, proj.path = ".", git.init = TRUE)
}
\arguments{
\item{proj.type}{type of project to create, based on the templates
found in \code{\link[prodigenr]{listTemplates}} function}

\item{proj.name}{name of the new project directory}

\item{proj.path}{location of where the new project will be created}

\item{git.init}{Logical, whether to initialize the new project
with git}
}
\value{
Creates a project directory with files and subdirectories
}
\description{
Create a project directory and associated files.
}
\details{
The main workhorse of the \pkg{prodigenr} package.  Creates a
project directory layout, with template files and helper scripts
such as a Makefile.
}
\examples{

\dontrun{
## Get a list of possible project templates
listTemplates('projects')
prodigen('testing', 'poster', 'dev/', FALSE)
prodigen('posterName', 'poster', 'path/to/dir/', FALSE)
prodigen('posterName', 'poster', 'path/to/dir/', TRUE)
prodigen('presentationName', 'slides', './', TRUE)
prodigen('abstractName', 'abstract', './', FALSE)
prodigen('manuscriptName', 'manuscript', './', FALSE)
}
}
\author{
Luke W. Johnston
}

