% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prodigen.R
\name{prodigen}
\alias{prodigen}
\title{Create a project directory based on a template}
\usage{
prodigen(proj.name, proj.path = ".", proj.type = "manuscript",
  git.init = TRUE)
}
\arguments{
\item{proj.name}{name of the new project directory}

\item{proj.path}{location of where the new project will be created}

\item{proj.type}{type of project to create, based on the templates
found in \code{\link[prodigenr]{listTemplates}} function}

\item{git.init}{Logical, whether to initialize the new project
with git}
}
\value{
Creates a project directory with files and subdirectories
}
\description{
Create a project directory and associated files.
}
\details{
The main workhorse of the \pkg{prodigenr} package.  Creates a
project directory layout, with template files and helper scripts
such as a Makefile.
}
\author{
Luke W. Johnston
}

