% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodigenr-package.R
\docType{package}
\name{prodigenr-package}
\alias{prodigenr}
\alias{prodigenr-package}
\title{prodigenr: Research Project Directory Generator}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create a project directory structure, along with typical files
    for that project.  This allows projects to be quickly and easily created,
    as well as for them to be standardized. Designed specifically with scientists
    in mind (mainly bio-medical researchers, but likely applies to other fields).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rostools/prodigenr}
  \item \url{http://rostools.github.io/prodigenr/}
  \item Report bugs at \url{https://github.com/rostools/prodigenr/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
