% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup a standardized folder and file structure for a research analysis project.}
\usage{
setup_project(path)
}
\arguments{
\item{path}{A path to a new directory.}
}
\value{
Project setup with folders and files necessary for a standard research project.
}
\description{
This starts the project by setting up a common folder and file infrastructure,
as well as adding some useful files to start the project. Also adds Git to the
project.
}
\examples{
\dontrun{
# Use a temporary location
new_proj_name <- fs::path_temp("DiabetesCancer")
setup_project(new_proj_name)
# After the new project opens up, add Git with:
setup_with_git()
}
}
